pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bleach>=6.3.0",
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "telegram>=0.0.1",
    "werkzeug>=3.1.4",
]
